address comments move numeral-type to numeral/numeral-type
